+++ src/onegov.pay/onegov/pay/models/payment.py
@@ -1,18 +1,16 @@
 from collections import namedtuple
 from decimal import Decimal
 from onegov.core.orm import Base
+from onegov.core.orm.abstract.associable import Associable
 from onegov.core.orm.mixins import ContentMixin
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
-from onegov.pay.utils import QueryChain
 from sqlalchemy import Column
 from sqlalchemy import Enum
-from sqlalchemy import Numeric
 from sqlalchemy import ForeignKey
+from sqlalchemy import Numeric
 from sqlalchemy import Text
 from sqlalchemy.ext.hybrid import hybrid_property
-from sqlalchemy.orm import joinedload
-from sqlalchemy.orm import object_session
 from uuid import uuid4
 
 
@@ -20,12 +18,10 @@ class RegisteredLink(namedtuple("RegisteredLink", ('cls', 'table', 'key'))):
     pass
 
 
-class Payment(Base, TimestampMixin, ContentMixin):
+class Payment(Base, TimestampMixin, ContentMixin, Associable):
     """ Represents a payment done through various means. """
 
     __tablename__ = 'payments'
-
-    registered_links = {}
 
     #: the public id of the payment
     id = Column(UUID, primary_key=True, default=uuid4)
@@ -81,42 +77,6 @@ class Payment(Base, TimestampMixin, ContentMixin):
 
         """
         return self.state == 'paid'
-
-    @staticmethod
-    def register_link(link_name, linked_class, association_table, table_key):
-        """ The :class:`~onegov.pay.models.payable.Payable` class registers
-        all back-referenes through this method. This is useful for two reasons:
-
-        1. We gain the ability to query all the linked records in one query.
-           This is hard otherwise as each ``Payable`` class leads to its own
-           association table which needs to be queried separately.
-
-        2. We are able to reset all created backreferences. This is necessary
-           during tests. SQLAlchemy keeps these references around and won't
-           let us re-register the same model multiple times (which outside
-           of tests is completely reasonable).
-
-        Note that this is a staticmethod instead of a classmethod because
-        we want all links to be managed through the actual ``Payment`` class,
-        not through any of its subclasses.
-
-        """
-        Payment.registered_links[link_name] = RegisteredLink(
-            cls=linked_class,
-            table=association_table,
-            key=table_key
-        )
-
-    @property
-    def links(self):
-        session = object_session(self)
-
-        return QueryChain(tuple(
-            session.query(link.cls)
-            .filter_by(payment=self)
-            .options(joinedload(link.cls.payment))
-            for link in self.registered_links.values()
-        ))
 
     @property
     def remote_url(self):