+++ src/onegov.pay/onegov/pay/utils.py
@@ -1,20 +1,6 @@
 from collections import namedtuple
 from decimal import Decimal
 from functools import total_ordering
-from sqlalchemy_utils import QueryChain as QueryChainBase
-
-
-class QueryChain(QueryChainBase):
-    """ Extends SQLAlchemy Utils' QueryChain with some extra methods. """
-
-    def slice(self, start, end):
-        return self[start:end]
-
-    def first(self):
-        return next((o for o in self), None)
-
-    def all(self):
-        return tuple(self)
 
 
 @total_ordering