+++ src/onegov.pay/onegov/pay/__init__.py
@@ -8,7 +8,7 @@ from onegov.pay.models import Payable, Payment, PaymentProvider
 from onegov.pay.collections import PaymentCollection, PayableCollection
 from onegov.pay.collections import PaymentProviderCollection
 from onegov.pay.integration import PayApp, process_payment
-from onegov.pay.utils import Price
+from onegov.pay.utils import Price, payments_association_table_for
 
 
 __all__ = (
@@ -23,5 +23,6 @@ __all__ = (
     'PaymentProvider',
     'PaymentProviderCollection',
     'Price',
-    'process_payment'
+    'process_payment',
+    'payments_association_table_for'
 )